+++ packages/stream_video_flutter/lib/src/call_screen/call_content/call_content.dart
@@ -181,6 +181,7 @@ class _StreamCallContentState extends State<StreamCallContent>
             call: call,
             participants: callState.callParticipants,
             layoutMode: ParticipantLayoutMode.pictureInPicture,
+            sort: widget.pictureInPictureConfiguration.sort,
           );
     }
 
@@ -196,10 +197,9 @@ class _StreamCallContentState extends State<StreamCallContent>
               width: 300,
               child: StreamPictureInPictureUiKitView(
                 call: call,
-                includeLocalParticipantVideo: widget
-                    .pictureInPictureConfiguration
-                    .iOSPiPConfiguration
-                    .includeLocalParticipantVideo,
+                configuration:
+                    widget.pictureInPictureConfiguration.iOSPiPConfiguration,
+                participantSort: widget.pictureInPictureConfiguration.sort,
               ),
             ),
           widget.callParticipantsBuilder?.call(